Move the downloads list files to "downloads" package
