+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/products/downloads/ProductDownloadsModule.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.products
+package com.woocommerce.android.ui.products.downloads
 
 import android.os.Bundle
 import androidx.lifecycle.ViewModel
@@ -6,6 +6,8 @@ import androidx.navigation.fragment.findNavController
 import androidx.savedstate.SavedStateRegistryOwner
 import com.woocommerce.android.R
 import com.woocommerce.android.di.ViewModelAssistedFactory
+import com.woocommerce.android.ui.products.ProductDetailViewModel
+import com.woocommerce.android.ui.products.ProductDetailViewModel.Factory
 import com.woocommerce.android.viewmodel.ViewModelKey
 import dagger.Binds
 import dagger.Module
@@ -13,7 +15,6 @@ import dagger.Provides
 import dagger.multibindings.IntoMap
 
 @Module
-
 abstract class ProductDownloadsModule {
     @Module
     companion object {
@@ -33,5 +34,5 @@ abstract class ProductDownloadsModule {
     @Binds
     @IntoMap
     @ViewModelKey(ProductDetailViewModel::class)
-    abstract fun bindFactory(factory: ProductDetailViewModel.Factory): ViewModelAssistedFactory<out ViewModel>
+    abstract fun bindFactory(factory: Factory): ViewModelAssistedFactory<out ViewModel>
 }
